refactor: share code between sign in and register
