+++ src/client/pages/Registration.tsx
@@ -3,34 +3,18 @@ import { QueryParams } from '@/shared/model/QueryParams';
 import { MainLayout } from '@/client/layouts/Main';
 import { MainForm } from '@/client/components/MainForm';
 import { EmailInput } from '@/client/components/EmailInput';
-import { TabType } from '@/client/components/Nav';
+import { generateSignInRegisterTabs } from '@/client/components/Nav';
 import { buildUrlWithQueryParams } from '@/shared/lib/routeUtils';
 import { Divider } from '@guardian/source-react-components-development-kitchen';
 import { SocialButtons } from '@/client/components/SocialButtons';
-import { css } from '@emotion/react';
-import { border, from, space } from '@guardian/source-foundations';
+import { socialButtonDivider } from '@/client/styles/Shared';
+import { usePageLoadOphanInteraction } from '@/client/lib/hooks/usePageLoadOphanInteraction';
 
 export type RegistrationProps = {
   email?: string;
   recaptchaSiteKey: string;
   queryParams: QueryParams;
 };
-
-const divider = css`
-  /* Undoes the negative margin */
-  margin-bottom: 0;
-  margin-top: ${space[4]}px;
-  ${from.mobileMedium} {
-    margin-top: ${space[6]}px;
-  }
-  :before,
-  :after {
-    content: '';
-    flex: 1 1;
-    border-bottom: 1px solid ${border.secondary};
-    margin: 8px;
-  }
-`;
 
 export const Registration = ({
   email,
@@ -46,20 +30,12 @@ export const Registration = ({
   const [recaptchaErrorContext, setRecaptchaErrorContext] =
     useState<ReactNode>(null);
 
-  const tabs: TabType[] = [
-    {
-      displayText: 'Sign in',
-      linkTo: '/signin',
-      queryParams: queryParams,
-      isActive: false,
-    },
-    {
-      displayText: 'Register',
-      queryParams: queryParams,
-      linkTo: '/register',
-      isActive: true,
-    },
-  ];
+  usePageLoadOphanInteraction(formTrackingName);
+
+  const tabs = generateSignInRegisterTabs({
+    queryParams,
+    isActive: 'register',
+  });
 
   return (
     <MainLayout
@@ -84,7 +60,7 @@ export const Registration = ({
       <Divider
         spaceAbove="loose"
         displayText="or continue with"
-        cssOverrides={divider}
+        cssOverrides={socialButtonDivider}
       />
       <SocialButtons queryParams={queryParams} />
     </MainLayout>
